Add provider-neutral frontend auth adapter
