+++ frontend/src/auth/NachetAuthProvider.tsx
@@ -1,12 +1,24 @@
-import React from "react";
-import { AuthApiBridge } from "./AuthApiBridge";
-import { NachetAuthContext } from "./NachetAuthContext";
+import { useEffect, type ReactNode } from "react";
+import { NachetAuthContext, useNachetAuth } from "./NachetAuthContext";
 import { MsalAuthProvider } from "./msal/MsalAuthProvider";
 import { OidcAuthProvider } from "./oidc/OidcAuthProvider";
+import { initializeApi } from "../common/api";
+import { errorLogger } from "../logging";
 
 export interface NachetAuthProviderProps {
   apiScopeClaim: string;
-  children: React.ReactNode;
+  children: ReactNode;
+}
+
+function AuthApiBridge() {
+  const { getAccessToken } = useNachetAuth();
+
+  useEffect(() => {
+    initializeApi(getAccessToken);
+    errorLogger.setTokenProvider(async () => getAccessToken());
+  }, [getAccessToken]);
+
+  return null;
 }
 
 export function NachetAuthProvider({
