+++ frontend/src/auth/oidc/OidcAuthProvider.test.tsx
@@ -1,7 +1,7 @@
 import { createContext, useContext } from "react";
 import { render, screen } from "@testing-library/react";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import type { NachetAuthContextValue } from "../types";
+import type { NachetAuthContextValue } from "../NachetAuthContext";
 import { OidcAuthProvider } from "./OidcAuthProvider";
 
 const mockOidcAuth = vi.hoisted(() => ({
