+++ frontend/src/root/body/body.tsx
@@ -667,7 +667,6 @@ const Body: React.FC<params> = (props) => {
             onClose={() => {
               /* Auth popup closes automatically when user is authenticated */
             }}
-            apiScopeClaim={apiScopeClaim}
           />
         )}
         {registrationModalOpen && (
